Reorders Components and remove containers
