+++ src/components/Builder/BuilderWrapper.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { DndProvider } from 'react-dnd';
 import TouchBackend from 'react-dnd-touch-backend';
-import Builder from '../containers/Builder';
+import { Builder } from './Builder';
 
 export const BuilderWrapper = () => {
   return (
