+++ src/components/Builder/ItemDroppable.js
@@ -2,19 +2,19 @@ import React from 'react';
 import classNames from "classnames";
 import { useDispatch, useSelector } from 'react-redux';
 import { useDrop } from 'react-dnd';
-import { addDraggedFood } from '../actions/newDiet/addDraggedFood';
-import { removeDraggedFood } from '../actions/newDiet/removeDraggedFood';
-import { changeFoodQuantity } from '../actions/newDiet/changeFoodCuantity';
-import { getFoodFromId } from '../selectors/foods/getFoodFromId';
-import { getNewDietFood } from '../selectors/newDiet/getNewDietFood';
-import { getNewDiet } from '../selectors/newDiet/getNewDiet';
-import { getFoods } from '../selectors/foods/getFoods';
-import { getMacrosPecent } from '../utils/getMacrosPecent';
-import { Pie } from '../components/Pie';
-import removeIcon from '../images/remove.svg';
-import { RealMacroQtty } from '../components/RealMacroQtty';
-import { getRealKCalQtty } from '../utils/getRealKCalQtty';
-import { getRealQtty } from '../utils/getRealQtty';
+import { addDraggedFood } from '../../actions/newDiet/addDraggedFood';
+import { removeDraggedFood } from '../../actions/newDiet/removeDraggedFood';
+import { changeFoodQuantity } from '../../actions/newDiet/changeFoodCuantity';
+import { getFoodFromId } from '../../selectors/foods/getFoodFromId';
+import { getNewDietFood } from '../../selectors/newDiet/getNewDietFood';
+import { getNewDiet } from '../../selectors/newDiet/getNewDiet';
+import { getFoods } from '../../selectors/foods/getFoods';
+import { getMacrosPecent } from '../../utils/getMacrosPecent';
+import { Pie } from '../Common/Pie';
+import removeIcon from '../../images/remove.svg';
+import { RealMacroQtty } from '../Common/RealMacroQtty';
+import { getRealKCalQtty } from '../../utils/getRealKCalQtty';
+import { getRealQtty } from '../../utils/getRealQtty';
 
 export const ItemDroppable = ({ foodCodes, meal }) => {
   const dispatch = useDispatch();
