+++ src/components/Common/BlankSlate.js
